Moved mergeFiles to it's seperate file
